major refactor of index.js, in prep for 3.x release.
