+++ test/usage.js
@@ -690,9 +690,7 @@ describe('usage tests', function () {
             r.should.have.property('errors');
             r.should.have.property('logs').with.length(1);
             r.should.have.property('exit').and.be.ok;
-            r.logs.join('\n').split(/\n+/).should.deep.equal([
-                '1.0.1'
-            ]);
+            r.logs[0].should.eql('1.0.1\n');
         });
     });
